chnaged vec3 from double to float

+++ app/src/main/java/com/example/raytracer/Vec3.kt
@@ -2,29 +2,29 @@ package com.example.raytracer
 
 import kotlin.math.sqrt
 
-data class Vec3(var x: Double = 0.0, var y: Double = 0.0, var z: Double = 0.0) {
-    inline var r: Double
+data class Vec3(var x: Float = 0.0f, var y: Float = 0.0f, var z: Float = 0.0f) {
+    inline var r: Float
         get() = x
         set(value) {
             x = value
         }
 
-    inline var g: Double
+    inline var g: Float
         get() = y
         set(value) {
             y = value
         }
 
-    inline var b: Double
+    inline var b: Float
         get() = z
         set(value) {
             z = value
         }
 
 
-    operator fun plus(v: Double) = Vec3(x + v, x + y, x + z)
+    operator fun plus(v: Float) = Vec3(x + v, x + y, x + z)
     operator fun plus(v: Vec3) = Vec3(x + v.x, y + v.y, z + v.z)
-    operator fun plusAssign(v: Double) {
+    operator fun plusAssign(v: Float) {
         this.x += v
         this.y += v
         this.z += z
@@ -36,9 +36,9 @@ data class Vec3(var x: Double = 0.0, var y: Double = 0.0, var z: Double = 0.0) {
         this.z += v.z
     }
 
-    operator fun minus(v: Double) = Vec3(x - v, x - y, x - z)
+    operator fun minus(v: Float) = Vec3(x - v, x - y, x - z)
     operator fun minus(v: Vec3) = Vec3(x - v.x, y - v.y, z - v.z)
-    operator fun minusAssign(v: Double) {
+    operator fun minusAssign(v: Float) {
         this.x -= v
         this.y -= v
         this.z -= z
@@ -50,32 +50,32 @@ data class Vec3(var x: Double = 0.0, var y: Double = 0.0, var z: Double = 0.0) {
         this.z -= v.z
     }
 
-    operator fun times(v: Double) = Vec3(x * v, y * v, z * v)
+    operator fun times(v: Float) = Vec3(x * v, y * v, z * v)
     operator fun times(v: Vec3) = Vec3(x * v.x, y * v.y, z * v.z)
 
-    operator fun timesAssign(v: Double) {
+    operator fun timesAssign(v: Float) {
         this.x *= v
         this.y *= v
         this.z *= v
     }
 
-    operator fun div(v: Double) = Vec3(x / v, y / v, z / v)
-    operator fun divAssign(v: Double) {
+    operator fun div(v: Float) = Vec3(x / v, y / v, z / v)
+    operator fun divAssign(v: Float) {
         this.x /= v
         this.y /= v
         this.z /= v
     }
 
-    fun length(): Double {
+    fun length(): Float {
         return sqrt(lengthSqr())
     }
 
-    fun lengthSqr(): Double {
+    fun lengthSqr(): Float {
         return x * x + y * y + z * z
     }
 }
 
-fun dot(u: Vec3, v: Vec3): Double {
+fun dot(u: Vec3, v: Vec3): Float {
     return u.x * v.x +
             u.y * v.y +
             u.z * v.z
